// src/app/api/car-logbook/model/logbookEntry.ts
/**
 * Car Logbook Service REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LogbookEntry { 
    /**
     * ID
     */
    id?: string;
    /**
     * ID
     */
    vehicle?: string;
    startDate?: string;
    endDate?: string;
    kmStart?: number;
    kmEnd?: number;
    trip?: number;
    route?: string;
    comment?: string;
}